-updated to Visual Studio 2017 -added CS1 read support -flagged manifest as hidpi aware -fixes for 9901 timer handling, found via CS1 in BASIC and EasyBug -fix for "windowshade" mode where Classic99 started as just a title bar -allow debugger to write more than 1 or 2 bytes at a time (to CPU, VDP or GROM) -display last accessed GROM address next to the hardware register -fix current directory changing when breakpoints were loaded or saved -allow an interrupt to trigger immediately even if there aren't quite enough cycles left -fix VDP scanlines to draw more correctly regardless of such things as overdrive and breakpoint -add 'unsupported operation' debug to the file handlers so they don't silently fail when it's deliberately not supported

+++ addons/mpd.cpp
@@ -66,7 +66,14 @@ unsigned char mpdGrom[19*8192];
 bool bMpdActive = false;				// enables the various hacks needed to emulate this
 int nCurrentBank[3] = { 0, 0 ,0 };		// which GROM is mapped into each of the 3 banks
 const Byte nDefaultConfigBytes[] = { 
-	MPD_BANK_994AGROM0, ~MPD_BANK_994AGROM0, MPD_BANK_994ABASIC1, ~MPD_BANK_994ABASIC1, MPD_BANK_994ABASIC2, ~MPD_BANK_994ABASIC2, 0xff, 0x00
+	(Byte)MPD_BANK_994AGROM0, 
+    (Byte)~MPD_BANK_994AGROM0, 
+    (Byte)MPD_BANK_994ABASIC1, 
+    (Byte)~MPD_BANK_994ABASIC1, 
+    (Byte)MPD_BANK_994ABASIC2, 
+    (Byte)~MPD_BANK_994ABASIC2, 
+    0xff, 
+    0x00
 };
 
 // these functions will be different on the AVR since it needs to reference EEPROM
